% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{fill_time_series_gaps}
\alias{fill_time_series_gaps}
\title{Fill gaps in time series data}
\usage{
fill_time_series_gaps(
  data,
  fields,
  time_field = "time",
  delta = "guess",
  max_gap = Inf,
  ...
)
}
\arguments{
\item{data}{Input data frame}

\item{fields}{Field names to interpolate}

\item{time_field}{Name of the time field in \code{data}}

\item{delta}{Time step of the output data. If the input data is regular,
this should match the input time step. The special value 'guess' will
pick the median time gap; this should not be used with irregular input data.}

\item{max_gap}{Maximum gap length to fill, in multiples of \code{delta}; if
set to \code{Inf}, all gaps will be filled}

\item{...}{Additional arguments passed to \code{\link{interpolate_time_series}},
which does the gap filling.}
}
\value{
A data frame with a time field and each of the fields named in
\code{fields}.
}
\description{
Fill gaps in time series data by interpolating between known values.
}
