% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{interpolate_time_series}
\alias{interpolate_time_series}
\title{Interpolate time series data}
\usage{
interpolate_time_series(
  data,
  fields,
  times,
  time_field = "time",
  method = c("approx", "spline"),
  ...
)
}
\arguments{
\item{data}{Input data frame}

\item{fields}{Field names to interpolate}

\item{times}{Times to interpolate at}

\item{time_field}{Name of the time field in \code{data}}

\item{method}{Interpolation method. Either \code{\link[stats]{approx}} or
\code{\link[stats]{spline}}.}

\item{...}{Additional arguments passed to the interpolation function.}
}
\value{
A data frame with a time field and each of the fields named in
\code{fields}.
}
\description{
Interpolate data to make a time series regular and to fill in missing gaps.
}
