% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parent-knn.R
\name{parent_knn}
\alias{parent_knn}
\title{Find the Nearest Neighbours Among Each Location's Parents}
\usage{
parent_knn(x, n_parents, leaf_size = 40)
}
\arguments{
\item{x}{A matrix or vector in which each row/element represents a point.}

\item{n_parents}{The number of parent nearest neighbours to find.}

\item{leaf_size}{Leaf size for the KD-tree algorithm used in nearest
neighbour search; the algorithm may run faster for different values.}
}
\value{
A matrix where the ith row contains the indices of the parent nearest
neighbours of the ith input point. The first column by convention is equal to
i, so the matrix has \code{n_parents + 1} columns. Some entries can be \code{NA}.
}
\description{
For each row in \code{x}, find the k-nearest neighbours among the earlier
rows.
}
\examples{
x <- matrix(rnorm(100), nrow = 50)
parents <- parent_knn(x, 5)

}
