% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptide.R
\name{ptide}
\alias{ptide}
\title{Specify a PTide model and fit it to data}
\usage{
ptide(
  formula,
  data,
  start_time,
  tidal_constituents = c("M2", "S2"),
  mean_function = list(mean = 0, precision = 1/100),
  covariance_functions = list(ptide_exponential_covariance_function(),
    ptide_quasi_periodic_covariance_function(), ptide_white_noise_covariance_function()),
  time_field = "time",
  ...
)
}
\arguments{
\item{formula}{A formula specifying the model; see \code{\link{ptide_model}}
for more information.}

\item{data}{A data frame containing the data to fit the model to.}

\item{start_time}{A \code{POSIXct} object specifying the start time of the
model; see \code{\link{ptide_model}} for more information.}

\item{tidal_constituents}{A vector of strings specifying the tidal
constituents to include in the model; see \code{\link{ptide_model}} for more
information.}

\item{mean_function}{A list specifying the prior for the mean function.
See \code{\link{ptide_model}} for more information.}

\item{covariance_functions}{A list of functions specifying the covariance
structure of the model. See \code{\link{ptide_model}} for more information.}

\item{time_field}{A string specifying the name of the field in the data that
represents time.}

\item{...}{Additional arguments passed to \code{\link{ptide_fit}}.}
}
\description{
Specify and fit a PTide model to data.
}
