% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ptide_covariance}
\alias{ptide_covariance}
\alias{ptide_cross_covariance}
\title{Calculate covariances between times using a PTide fit}
\usage{
ptide_covariance(x, fit, include_noise)

ptide_cross_covariance(x1, x2, fit)
}
\arguments{
\item{x}{A vector of times}

\item{fit}{A PTide fit created using \code{\link{ptide_fit}}}

\item{include_noise}{A vector of 0s and 1s indicating whether to include
the white noise term for each time point.}

\item{x1}{The first vector of times for \code{ptide_cross_covariance}}

\item{x2}{A second vector of times for \code{ptide_cross_covariance}}
}
\value{
For \code{ptide_covariance}, a 4-D array of cross-covariances with
dimensions n x n x d x d, where n is the length of \code{x} and d is the
number of data dimensions. For \code{ptide_cross_covariance}, a 4-D array
of cross-covariances with dimensions n1 x n2 x d x d, where n1 and n2 are
the lengths of \code{x1} and \code{x2}.
}
\description{
Calculate the covariances between times using a PTide fit.
}
\section{Functions}{
\itemize{
\item \code{ptide_cross_covariance()}: PTide cross covariance

}}
