% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance-functions.R
\name{ptide_covariance_function}
\alias{ptide_covariance_function}
\alias{ptide_exponential_covariance_function}
\alias{ptide_matern32_covariance_function}
\alias{ptide_matern52_covariance_function}
\alias{ptide_periodic_covariance_function}
\alias{ptide_quasi_periodic_covariance_function}
\alias{ptide_white_noise_covariance_function}
\title{Covariance functions}
\usage{
ptide_exponential_covariance_function(
  sigma_squared_a = "empirical",
  sigma_squared_b = "empirical",
  L_shape = 1.1,
  ell_a = 0.1,
  ell_b = 0.1,
  length_dilation = 1
)

ptide_matern32_covariance_function(
  sigma_squared_a = "empirical",
  sigma_squared_b = "empirical",
  L_shape = 1.1,
  ell_a = 0.1,
  ell_b = 0.1,
  length_dilation = 1
)

ptide_matern52_covariance_function(
  sigma_squared_a = "empirical",
  sigma_squared_b = "empirical",
  L_shape = 1.1,
  ell_a = 0.1,
  ell_b = 0.1,
  length_dilation = 1
)

ptide_periodic_covariance_function(
  period = "M2",
  sigma_squared_a = "empirical",
  sigma_squared_b = "empirical",
  L_shape = 1.1,
  ell_a = 0.1,
  ell_b = 0.1,
  length_dilation = 1
)

ptide_quasi_periodic_covariance_function(
  period = "M2",
  sigma_squared_a = "empirical",
  sigma_squared_b = "empirical",
  L_shape = 1.1,
  ell_periodic_a = 0.1,
  ell_periodic_b = 0.1,
  ell_decay_a = 0.1,
  ell_decay_b = 0.1,
  length_dilation = 1
)

ptide_white_noise_covariance_function(
  sigma_squared_a = "empirical",
  sigma_squared_b = "empirical",
  L_shape = 1.1
)
}
\arguments{
\item{sigma_squared_a}{The shape parameter of inverse-gamma prior on sigma
squared. If set to 'empirical', this will be based on the empirical variance
of the data.}

\item{sigma_squared_b}{The rate parameter of inverse-gamma prior on sigma
squared.}

\item{L_shape}{The parameter of the LKJ prior for the correlation matrix C of
the covariance function.}

\item{ell_a}{The shape parameter of the inverse-gamma prior on the length
scale parameter, \eqn{\ell}.}

\item{ell_b}{The scale parameter of the inverse-gamma prior on the length
scale parameter, \eqn{\ell}.}

\item{length_dilation}{Used to multiplicatively scale the hours in the
covariance calculation, intended to help the user shift the value of the
length scale parameters to approximately one.}

\item{period}{The period of the periodic portion of a periodic or
quasi-periodic covariance function, in hours. Can also be given as a string,
in which case the period corresponds to one of the tidal frequencies given
in \code{\link{tidal_constituents}}.}

\item{ell_periodic_a}{The shape parameter of the inverse-gamma prior on the
periodic length scale parameter, ell_periodic, of the quasi-periodic
covariance function.}

\item{ell_periodic_b}{The rate parameter of the inverse-gamma prior on the
periodic length scale parameter, ell_periodic, of the quasi-periodic
covariance function.}

\item{ell_decay_a}{The shape parameter of the inverse-gamma prior on the
decay length scale parameter, ell_decay, of the quasi-periodic covariance
function.}

\item{ell_decay_b}{The rate parameter of the inverse-gamma prior on the decay
length scale parameter, ell_decay, of the quasi-periodic covariance function.}
}
\description{
These functions are used to specify covariance functions for use with
\code{\link{ptide}} and \code{\link{ptide_model}}.
}
\details{
The covariance functions in this package have a separable structure, with
\eqn{K(t, t') = k(t, t')S} a \eqn{d \times d} matrix, where d is the
dimension of the process, \eqn{k(t, t')} is a univariate covariance function,
and S is a \eqn{d \times d} covariance matrix. Internally, S is parameterised
as diag(sigma) \%\emph{\% C \%}\% diag(sigma), where sigma is a \eqn{d \times 1}
vector of standard deviations, and C is a correlation matrix. The univariate
covariate functions available are described below. All length scale
parameters are given inverse gamma priors, sigma^2 is given an inverse gamma
prior, and C is given the LKJ prior.
}
\section{Functions}{
\itemize{
\item \code{ptide_exponential_covariance_function()}: Exponential covariance function

\item \code{ptide_matern32_covariance_function()}: Matern 3/2 covariance function

\item \code{ptide_matern52_covariance_function()}: Matern 5/2 covariance function

\item \code{ptide_periodic_covariance_function()}: Periodic covariance function

\item \code{ptide_quasi_periodic_covariance_function()}: Quasi-periodic covariance function

\item \code{ptide_white_noise_covariance_function()}: White noise covariance function

}}
\section{Exponential}{

The exponential covariance function
(\code{ptide_exponential_covariance_function}) has the expression
\deqn{k(t, t') = \exp\left( -\frac{|t - t'|}{\ell / d} \right),}
where \eqn{\ell} is the length scale parameter, and d is a
user-specified length dilation (called \code{length_dilation} in the
parameters).
}

\section{Matern 3/2 and 5/2}{

The Matern 3/2 and 5/2 covariance functions
(\code{ptide_matern32_covariance_function} and
\code{ptide_matern52_covariance_function}) has the expression
\deqn{
  k(t, t') = \frac{
    2^{1 - \nu}}{\Gamma(\nu)
  }(\sqrt{2\nu}|t - t'| / \ell)^\nu K_\nu(\sqrt{2 \nu} |t - t'| / \ell),}
where \eqn{\nu = 3/2} or \eqn{5 / 2}, \eqn{\ell} is the length scale
parameter, and d is a user-specified length dilation.
}

\section{Squared exponential}{

The squared exponential covariance function
(\code{ptide_squared_exponential_covariance_function} in this package has
the expression
\deqn{k(t, t') = \exp\left( -\frac{(t - t')^2}{2(\ell / d)^2} \right),}
where \eqn{\ell} is the length scale parameter, and d is a user-specified
length dilation.
}

\section{Periodic}{

The periodic covariance function (\code{ptide_periodic_covariance_function}
in this package has the expression
\deqn{k(t, t')
  = \exp\left( -\frac{2\sin^2(\pi|t - t'| / p)}{\ell^2} \right),}
where \eqn{\ell} is the length scale parameter, and \eqn{p} is the period of
the covariance function.
}

\section{Quasi-periodic}{

The quasi-periodic covariance function
(\code{ptide_quasi_periodic_covariance_function} in this package has the
expression
\deqn{k(t, t') =
    \exp\left( -\frac{2\sin^2(\pi|t - t'| / p)}{\ell_p^2} \right)
    \exp\left( -\frac{(t - t')^2}{2(\ell_d / d)^2} \right),
}
where \eqn{\ell_p} is the periodic length scale parameter, \eqn{p} is the
period of the covariance function, \eqn{\ell_d} is the decay length scale,
and \eqn{d} is the length dilation, as described in the section on the
squared exponential covariance function.
}

\section{White noise}{

The white noise covariance function
(\code{ptide_white_noise_covariance_function} in this package has the
expression
\deqn{k(t, t') = I(t = t'),}
where \eqn{I(\cdot)} is an indicator function.
}

