% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-spectrum.R
\name{ptide_error_spectrum}
\alias{ptide_error_spectrum}
\title{Error spectrum}
\usage{
ptide_error_spectrum(
  fit,
  frequency_range = c(0, 1),
  delta = 0.1,
  lag_max = 500,
  plot = FALSE,
  ...
)
}
\arguments{
\item{fit}{A ptide fit object made using \code{\link{ptide_fit}}.}

\item{frequency_range}{A two-element vector giving the range of frequencies
to return. These are measured in cycles per hour.}

\item{delta}{The time step used in the numerical integration. This should be
made as small as possible subject to computing time.}

\item{lag_max}{The maximum lag to use in the numerical integration. This
should be large enough that the autocovariance is essentially zero at this
lag.}

\item{plot}{Whether to plot the spectrum.}

\item{...}{Additional arguments passed to \code{\link{plot}}.}
}
\value{
An n by d by d array. The first dimension corresponds to the
frequencies. The second and third dimensions correspond to the rows and
columns of the spectral matrix. The frequencies are stored as an attribute
named \code{frequencies}.
}
\description{
Calculate (and optionally plot) the spectrum of the error process for a ptide
fit. The spectrum is calculated through numerical integration based on the
discrete Fourier transform. The estimated spectral density matrix is returned
for a range of frequencies determined by \code{frequency_range},
\code{delta}, and \code{lag_max}.
}
\details{
The accuracy of the returned spectrum depends on both the \code{delta} and
\code{lag_max} arguments. These should be adjusted (made smaller for
\code{delta} and larger for \code{lag_max}) until the spectrum converges.
}
