% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{ptide_fit}
\alias{ptide_fit}
\alias{print.ptide_fit}
\title{Fit a PTide model to data}
\usage{
ptide_fit(
  model,
  data,
  vecchia_configuration = list(n_parents = 256L, n_responses = 256L),
  max_attempts = 20,
  best_of = 10,
  grain_size = "auto",
  threads = getOption("ptide.threads", -1L),
  ...
)

\method{print}{ptide_fit}(x, ...)
}
\arguments{
\item{model}{A PTide model specified with \code{\link{ptide_model}}.}

\item{data}{A data frame containing the data to fit the model to.}

\item{vecchia_configuration}{A list containing the configuration for the
Vecchia approximation. The list must contain two elements: \code{n_parents}
(number of parents) and \code{n_responses} (number of responses). For the
approximation, the time series is split into blocks of length
\code{n_responses}, and each block conditions on the previous
\code{n_parents} time points.}

\item{max_attempts}{Maximum number of optimisation attempts.}

\item{best_of}{Number of times to run the optimisation algorithm. The
algorithm is run repeatedly with different starting values in order to help
find the global posterior mode.}

\item{grain_size}{Used to control the parallelisation. If set to 'auto',
this will attempted to maximise the use of threads}

\item{threads}{Number of threads to use for parallelisation, taken from
\code{getOption('ptide.threads')} by default, which is itself set to -1 by
default. The special value -1 picks a number of threads based on the number
of cores in the system).}

\item{...}{Additional parameters to be passed to
\code{\link{stan_optimizing_best_of}}}

\item{x}{Object to be printed in \code{print.ptide_fit}.}
}
\value{
A PTide fit object (a list) with the following entries:
\itemize{
\item \code{model}: The PTide model object.
\item \code{covariance_functions}: A list of covariance function objects,
one for each covariance function in the model, with the fitted
parameters.
\item \code{beta_hat}: The fitted mean function parameters.
\item \code{optimization_result}: The raw output of the optimisation
procedure.
}
}
\description{
Fits a PTide model specified with \code{\link{ptide_model}} to data.
}
\details{
The computational time required for the fit depends on the number of
observations, the choice of Vecchia approximation (see the
\code{vecchia_configuration} parameter below), and whether the observed
times are regular. The last point is particularly important: if the observed
optimisation routine will run much faster.
}
\section{Methods (by generic)}{
\itemize{
\item \code{print(ptide_fit)}: Print a fit to screen

}}
