% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{ptide_model}
\alias{ptide_model}
\title{Specify a PTide model}
\usage{
ptide_model(
  formula,
  start_time,
  tidal_constituents = c("M2", "S2"),
  mean_function = list(mean = 0, precision = 1/100),
  covariance_functions = list(ptide_exponential_covariance_function(),
    ptide_quasi_periodic_covariance_function(), ptide_white_noise_covariance_function()),
  time_field = "time"
)
}
\arguments{
\item{formula}{An object of class \code{\link[stats]{formula}} specifying the
model to be fitted. This formula includes the response variable or variables
and potentially external covariates. If more than one response variable is
specified (using \link{cbind}, the model is a multivariate model.}

\item{start_time}{A POSIXct object representing the start time of the time
series. This is used to align the model with the time field of the data. If
not provided, a default of '2020-01-01 00:00:00' is used.}

\item{tidal_constituents}{A character vector specifying the tidal
constituents to be included in the model. Default constituents are 'M2' and
'S2'.}

\item{mean_function}{A list specifying the prior for the coefficients of the
mean function of the model. The list should contain two elements: 'mean'
(the prior mean), either a scalar or a vector, and 'precision' (the prior
Default values are 0 for mean and 1/100 for precision.}

\item{covariance_functions}{A list of functions specifying the covariance
structure of the model. Each function in the list should be one of the
covariance functions defined in the package (like
\code{\link{ptide_exponential_covariance_function}}). By default,
exponential, quasi-periodic, and white noise covariance functions are
included.}

\item{time_field}{A string specifying the name of the field in the data that}
}
\value{
An object of class 'ptide_model' that represents the specified PTide
model.
}
\description{
Describe a PTide model, including response variable, external covariates,
tidal constituents, and covariance functions. This function allows the user
to specify the structure of a PTide model, including its formula, start time,
tidal constituents, mean function, covariance functions, and the time field.
}
\examples{
# Bivariate response
ptide_model(
  formula = cbind(easting, northing) ~ 1,
  start_time = lubridate::ymd_hms('2023-01-01 00:00:00'),
  tidal_constituents = c('M2', 'S2', 'K1')
)

}
