% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{ptide_predict}
\alias{ptide_predict}
\title{Predict using PTide}
\usage{
ptide_predict(
  fit,
  predicted_df,
  observed_df,
  include_sd = FALSE,
  include_marginal_covariance = FALSE,
  include_covariance = FALSE,
  include_white_noise = TRUE,
  n_samples = 0,
  method = c("standard", "general"),
  vecchia_configuration = list(n_parents = 256L, n_responses = 256L, n_nearest = 256L),
  cache,
  cache_resolution = 60L
)
}
\arguments{
\item{fit}{A PTide fit object made using \code{\link{ptide_fit}}.}

\item{predicted_df}{A data frame containing the times to be predicted, and
the values of any covariates used in the fit.}

\item{observed_df}{A data frame containing the observed data that will be
used for the prediction.}

\item{include_sd}{Whether to include the standard deviation of the
prediction in the output.}

\item{include_marginal_covariance}{Whether to include the marginal
covariance of the prediction in the output.}

\item{include_covariance}{Whether to include the full covariance of the
prediction in the output.}

\item{include_white_noise}{Whether to include the white noise kernel in the
prediction variance and the samples. This is only relevant if the white noise
kernel was included in the fit.}

\item{n_samples}{The number of samples to draw from the prediction
distribution. If set to 0, no samples will be drawn.}

\item{method}{The method to use for the prediction. See the methods section
below.}

\item{vecchia_configuration}{A list containing the configuration for the
Vecchia approximation. See \code{\link{ptide_fit}} for documentation on this
field; in general, larger values are better, but take longer to compute.
The \code{n_nearest} entry is used for the standard method when the
prediction times are not in the future relative to the observed data.}

\item{cache}{A cache object satisfying the semantics of the \code{cachem}
package, such as \code{\link[cachem]{cache_mem}}. This can be used to save
the results of expensive computations between calls to \code{ptide_predict}.
If omitted, no cache is used.}

\item{cache_resolution}{The resolution of the cache, in seconds. This is
only relevant if \code{cache} is specified. Covariances over differences
smaller than this resolution are considered to be identical.}
}
\value{
A list containing the predicted values. The list will contain the
following elements:
\itemize{
\item \code{fitted}: The fitted values, which is equal to the mean function
evaluated at the prediction times.
\item \code{mean}: The posterior mean of the prediction distribution.
\item \code{sd}: The posterior standard deviation of the prediction
distribution. This will only be included if \code{include_sd} is set to
\code{TRUE}.
\item \code{marginal_covariance}: The posterior covariance matrix at each
time point that is predicted. This will only be included if
\code{include_marginal_covariance} is set to \code{TRUE}.
\item \code{covariance}: The posterior covariance matrix between all
predicted time points. This will only be included if
\code{include_covariance} is set to \code{TRUE}.
\item \code{samples}: Samples from the posterior distribution. This will
only be included if \code{n_samples} is greater than 0.
}
}
\description{
Perform predictions using a PTide fit. The predictions are made at chosen
times and will be conditioned on the observed data and the covariance
parameters estimated by \code{\link{ptide_fit}}. The covariate coefficients
and the coefficients of the tidal harmonics are re-estimated using the
observed data.
}
\details{
The computational time required for the prediction depends on the number of
observations, the number of predicted time points, the choice of Vecchia
approximation (see the \code{vecchia_configuration} parameter below), and
whether the observed times are regular. The last point is particularly
important: if the observed times are regular, many matrix operations can be
precomputed and the prediction will be much faster.
}
\section{Methods}{

Two methods are available:
\itemize{
\item \code{standard}: This method corresponds to the standard Vecchia
approximation. This is fast when the prediction times are in the
future, and reasonably fast when some are in the past. It is very
accurate if the number of parents in the Vecchia configuration is high
and when the white noise component of the covariance structure is
relatively small.
\item \code{general}: This method corresponds to the general Vecchia
approximation of Katzfuss et. al (2020). This is much slower than the
standard method but may be more accurate in some situations.
}
}

