% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{run_validation}
\alias{run_validation}
\title{Perform time series cross-validation for PTide}
\usage{
run_validation(
  fit,
  data,
  start_index,
  horizons,
  skip = 1L,
  last_n = Inf,
  include_sd = FALSE,
  include_marginal_correlation = FALSE,
  cores = 1L,
  cache = cachem::cache_mem(),
  ...
)
}
\arguments{
\item{fit}{A PTide fit object as returned by \code{\link{ptide_fit}}.}

\item{data}{A data frame containing the data to be used for the validation.}

\item{start_index}{The row number of the first observation from which to
start the validation.}

\item{horizons}{A vector of time horizons for which to make predictions.
Should be a period of time created using \code{\link{lubridate::hours}} or
similar.}

\item{skip}{The number of observations to skip between each prediction.}

\item{last_n}{The number of observations to condition on for the prediction.
If \code{Inf}, conditions on all previous observations.}

\item{include_sd}{Whether to include the standard deviation of the forecast
in the output.}

\item{include_marginal_correlation}{Whether to include the marginal
correlation of the forecast in the output.}

\item{cores}{The number of cores to use for parallelisation. If \code{1},
parallelisation is disabled.}

\item{cache}{A cache object satisfying the semantics of the \code{cachem}
package, such as \code{\link[cachem]{cache_mem}}. This is passed to
\code{\link{ptide_predict}}.}

\item{...}{Additional arguments to be passed to \code{\link{ptide_predict}}.}
}
\value{
A data frame containing the predictions alongside the input data.
}
\description{
Perform time series cross-validatiom for a PTide fit. Given a dataset,
this method will perform a series of predictions by advancing the
observed data by \code{skip} observation at a time.
}
