% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbb.R
\name{set_tbb_threads}
\alias{set_tbb_threads}
\alias{get_tbb_threads}
\title{Get and set number of TBB threads}
\usage{
set_tbb_threads(n)

get_tbb_threads()
}
\arguments{
\item{n}{Number of threads}
}
\value{
The number of threads prior to the call
}
\description{
Get and set the number of threads used by Intel's Threading Building Blocks
(TBB) library, used within Stan for parallelisation.
}
\section{Functions}{
\itemize{
\item \code{get_tbb_threads()}: Get number of TBB threads

}}
