% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimisation.R
\name{stan_nlminb}
\alias{stan_nlminb}
\title{Find Optimal Parameters for a Stan Model Using nlminb}
\usage{
stan_nlminb(
  object,
  data,
  trace = 2,
  control = list(iter.max = 2000, eval.max = 10000, trace = trace),
  initial_value_attempts = 10,
  theta_initial,
  ...
)
}
\arguments{
\item{object}{A Stan model object.}

\item{data}{A list containing the data to be used in the Stan model.}

\item{trace}{How frequently to print the current iteration (every \code{trace}
iterations). Default is 2.}

\item{control}{A list of control parameters passed to
\code{\link{nlminb}}.}

\item{initial_value_attempts}{The number of attempts to find a valid initial
value for the optimisation. Default is 10.}

\item{theta_initial}{The initial value for the parameters. If not supplied,
a random value will be used.}

\item{...}{Additional arguments passed to \code{\link{nlminb}}.}
}
\value{
A list containing the following:
\itemize{
\item \code{value}: The optimized value of the objective function.
\item \code{par}: The optimized parameters in the original parameter space.
\item \code{iterations}: The number of iterations taken by the algorithm.
\item \code{evaluations}: The number of function evaluations.
\item \code{convergence}: The convergence code returned by \code{\link{nlminb}}.
\item \code{message}: The message returned by \code{\link{nlminb}}.
}
}
\description{
This function optimizes the parameters of a given Stan model using
\code{\link{nlminb}}. This function internally uses the
\code{\link[rstan]{log_prob}} and \code{\link[rstan]{grad_log_prob}}
functions from the \pkg{rstan} package.
}
\examples{
\dontrun{
  # Define a Stan model (assuming `model_code` contains your Stan model)
  stan_model <- rstan::stan_model(model_code = model_code)
  # Define data (assuming `data_list` contains your data)
  fit <- stan_nlminb(stan_model, data = data_list)
}

}
\seealso{
\code{\link{nlminb}}
}
