% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidal-harmonics.R
\docType{data}
\name{tidal_constituents}
\alias{tidal_constituents}
\title{Tidal constituents}
\format{
A data.frame
}
\usage{
tidal_constituents
}
\description{
A data.frame of tidal constituents giving names, frequencies in hours, and
Doodson numbers. Copied from \link[oce]{tidedata} in the oce package, who in
which in turn derives it from Foreman (1979).
}
\keyword{datasets}
