% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidal-harmonics.R
\name{tidal_harmonic_design_matrix}
\alias{tidal_harmonic_design_matrix}
\title{Create a tidal harmonic design matrix}
\usage{
tidal_harmonic_design_matrix(
  times,
  hours,
  reference_time = min(times),
  constituent_names = c("M2", "S2"),
  intercept = FALSE,
  trend = FALSE
)
}
\arguments{
\item{times}{The times at which to create the harmonics. A vector of date
time objects for which difftime is defined.}

\item{hours}{A numeric vector of times at which to create the harmonics,
in hours. Ignored if \code{times} are provided.}

\item{reference_time}{The reference time (time zero) for the harmonics.
Ignored if \code{hours} is provided}

\item{constituent_names}{The names of the tidal harmonics to include. See
\code{\link[ptide]{tidal_constituents}}.}

\item{intercept}{Whether to include an intercept term as a column of ones.}

\item{trend}{Whether to include a trend column in hours.}
}
\value{
A data.frame with the chosen frequencies as columns.
}
\description{
Creates a data.frame with harmonics at the chosen frequencies.
}
\examples{
design_matrix <- tidal_harmonic_design_matrix(hours = 1 : 30)
print(design_matrix)
}
