% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidal-harmonics.R
\name{tidal_period}
\alias{tidal_period}
\title{Tidal period}
\usage{
tidal_period(period)
}
\arguments{
\item{period}{A tidal constituent name, or a numeric value in hours.}
}
\value{
The period in hours.
}
\description{
Converts a tidal constituent name to its period in hours.
}
